transcript_image/requirements.txt
numpy==2.0.2
opencv-python==4.11.0.86
pillow==11.1.0
streamlit==1.44.1